\name{mrna}
\docType{data}
\alias{mrna}
\title{mRNA sequence of gene in ProbeLevelSet}
\description{
This is the mRNA sequence for fibronectin 1 isoform 2 preproprotein, which is used in the exampleProbeLevelSet.


  }
\usage{data(exampleProbeLevelSet)}
\format{
	It has been read from FASTA format using the \link[Biostrings:readDNAStringSet]{readDNAStringSet} function.
}
\author{Lasse Folkersen}
\keyword{datasets}