\name{Nondocumented-objects}
\alias{Nondocumented-objects}

\title{Here goes every undocumented function.}
\description{Every function without help page is redirected here. Either because it is not 
intended to have a help page or because it is a bug. If you think it is the 
later please contact the package maintainer.

}

\author{Lasse Folkersen and Diego Diez}
\keyword{documentation}
\keyword{internal}
