\name{ProbeLevelSet-class}
\docType{class}
\alias{ProbeLevelSet}
\alias{ProbeLevelSet-class}


\title{Class ProbeLevelSet}
\description{This is the storage class for Probe Level data with attached probe annotation}
\section{Class definition}{
  The class is derived from class ExpressionSet in the Biobase package.}

\author{Diego Diez}

\keyword{classes}